skills/arxiv/scripts/pyproject.toml
[project]
name = "arxiv-scripts"
version = "0.1.0"
description = "arXiv paper search and download tool"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.1.0",
    "markitdown[all]>=0.1.0",
    "python-dateutil>=2.8.2",
    "arxiv-to-prompt>=0.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["arxiv_tool.py", "arxiv_fetcher.py"]
